+++ Events/DockingGrantedEvent.cs
@@ -1,4 +1,5 @@
-﻿using Newtonsoft.Json;
+﻿using EddiDataDefinitions;
+using Newtonsoft.Json;
 using System;
 using System.Collections.Generic;
 
@@ -14,15 +15,18 @@ namespace EddiEvents
         static DockingGrantedEvent()
         {
             VARIABLES.Add("station", "The station at which the commander has been granted docking");
-            VARIABLES.Add("stationtype", "The model / type of the station at which the commander has been granted docking");
+            VARIABLES.Add("stationtype", "The localized model / type of the station at which the commander has been granted docking");
+            VARIABLES.Add("stationDefinition", "The model / type of the station at which the commander has been granted docking (this is an object)");
             VARIABLES.Add("landingpad", "The landing pad at which the commander has been granted docking");
         }
 
         [JsonProperty("station")]
         public string station { get; private set; }
 
-        [JsonProperty("stationtype")]
-        public string stationtype { get; private set; }
+        [JsonProperty("stationType")]
+        public StationModels stationDefinition { get; private set; }
+
+        public string stationtype => stationDefinition.localizedName;
 
         [JsonProperty("landingpad")]
         public int landingpad { get; private set; }
@@ -30,7 +34,7 @@ namespace EddiEvents
         public DockingGrantedEvent(DateTime timestamp, string station, string stationType, int landingpad) : base(timestamp, NAME)
         {
             this.station = station;
-            this.stationtype = stationType;
+            this.stationDefinition = StationModels.FromEDName(stationType);
             this.landingpad = landingpad;
         }
     }