+++ Events/DockingTimedOutEvent.cs
@@ -1,4 +1,5 @@
-﻿using Newtonsoft.Json;
+﻿using EddiDataDefinitions;
+using Newtonsoft.Json;
 using System;
 using System.Collections.Generic;
 
@@ -14,19 +15,22 @@ namespace EddiEvents
         static DockingTimedOutEvent()
         {
             VARIABLES.Add("station", "The station at which the docking request has timed out");
-            VARIABLES.Add("stationtype", "The model / type of the station at which the commander has been granted docking");
+            VARIABLES.Add("stationtype", "The localized model / type of the station at which docking has timed out");
+            VARIABLES.Add("stationDefinition", "The model / type of the station at which docking has timed out (this is an object)");
         }
 
         [JsonProperty("station")]
         public string station { get; private set; }
 
-        [JsonProperty("stationtype")]
-        public string stationtype { get; private set; }
+        [JsonProperty("stationType")]
+        public StationModels stationDefinition { get; private set; }
+
+        public string stationtype => stationDefinition.localizedName;
 
         public DockingTimedOutEvent(DateTime timestamp, string station, string stationType) : base(timestamp, NAME)
         {
             this.station = station;
-            this.stationtype = stationType;
+            this.stationDefinition = StationModels.FromEDName(stationType);
         }
     }
 }